modifying page pull so it gets static product initially then supplements with dynamic inventory data, also moved fetch products and related products to use-layout-data hooks file
