+++ src/pages/products/[handle].tsx
@@ -1,4 +1,3 @@
-import { medusaClient } from "@lib/config"
 import { IS_BROWSER } from "@lib/constants"
 import { getProductHandles } from "@lib/util/get-product-handles"
 import Head from "@modules/common/components/head"
@@ -11,45 +10,18 @@ import { useRouter } from "next/router"
 import { ParsedUrlQuery } from "querystring"
 import { ReactElement } from "react"
 import { NextPageWithLayout, PrefetchedPageProps } from "types/global"
-import { fetchCollectionData, fetchRegionsData, fetchCategoryData, formatProducts } from "@lib/hooks/use-layout-data"
-import { Region } from "@medusajs/medusa"
-import { PricedProduct } from "@medusajs/medusa/dist/types/pricing"
-import { ProductPreviewType } from "types/global"
+import { fetchCollectionData, fetchRegionsData, fetchCategoryData, fetchProduct, fetchRelatedProducts } from "@lib/hooks/use-layout-data"
 
 interface Params extends ParsedUrlQuery {
   handle: string
-}
-
-const fetchProduct = async (handle: string) => {
-  return await medusaClient.products
-    .list({ handle })
-    .then(({ products }) => products[0])
-}
-
-export const fetchRelatedProducts = async (
-region: Region,
-handle: string
-): Promise<ProductPreviewType[]> => {
-  const products = await medusaClient.products
-    .list({
-      region_id: region.id,
-      is_giftcard: false,
-      limit: 5,
-    })
-    .then(({ products }) => products)
-    .catch((_) => [] as PricedProduct[])
-
-  // filter out current product if it exists in the array and ensure 4 products returned
-  const filteredProducts = products.filter((product) => product.handle !== handle).slice(0, 4)
-
-  return formatProducts(filteredProducts, region)
 }
 
 const ProductPage: NextPageWithLayout<PrefetchedPageProps> = ({ notFound }) => {
   const { query, isFallback, replace } = useRouter()
   const handle = typeof query.handle === "string" ? query.handle : ""
 
-  const { data, isError, isLoading, isSuccess } = useQuery(
+  // get product statically (without inventory data)
+  const { data: product, isError, isLoading, isSuccess } = useQuery(
     [`get_product`, handle],
     () => fetchProduct(handle),
     {
@@ -57,6 +29,16 @@ const ProductPage: NextPageWithLayout<PrefetchedPageProps> = ({ notFound }) => {
       keepPreviousData: true,
     }
   )
+
+  // grab product dynamically (with inventory)
+  const { data: productWithInventory } = useQuery(
+    [`product_inventory`, handle],
+    () => fetchProduct(handle, true),
+    {
+      enabled: handle.length > 0,
+      keepPreviousData: true,
+    }
+  )  
 
   if (notFound) {
     if (IS_BROWSER) {
@@ -66,7 +48,7 @@ const ProductPage: NextPageWithLayout<PrefetchedPageProps> = ({ notFound }) => {
     return <SkeletonProductPage />
   }
 
-  if (isFallback || isLoading || !data) {
+  if (isFallback || isLoading || !product) {
     return <SkeletonProductPage />
   }
 
@@ -74,15 +56,16 @@ const ProductPage: NextPageWithLayout<PrefetchedPageProps> = ({ notFound }) => {
     replace("/404")
   }
 
+  // use the static product initially and then replace with dynamic once received
   if (isSuccess) {
     return (
       <>
         <Head
-          description={data.description}
-          title={data.title}
-          image={data.thumbnail}
+          description={product.description}
+          title={product.title}
+          image={product.thumbnail}
         />
-        <ProductTemplate product={data} />
+        <ProductTemplate product={productWithInventory || product} />
       </>
     )
   }
